chore: Deprecate legacy app management helpers
